requirements.txt
# KUBE-AI Dependencies (MindSpore)
mindspore>=2.0.0
numpy>=1.21.0
Pillow>=8.3.0
requests>=2.25.0
matplotlib>=3.5.0

# API (optional)
flask>=2.3.0

# Video processing (optional)
opencv-python>=4.5.0

# Dataset download (optional)
roboflow>=1.0.0
kaggle>=1.5.0
